Apply theme to Gradient component - Create a separate Loading scene for the EdgeCoreManager - Use Scenewrapper on TransactionDetails as the gradient background - Remove old Gradient component and TransactionDetails unused styles
