+++ src/modules/UI/components/Gradient/Gradient.ui.js
@@ -1,38 +1,47 @@
 // @flow
 
 import type { Node } from 'react'
-import React, { Component } from 'react'
+import React, { PureComponent } from 'react'
 import { StyleSheet } from 'react-native'
 import LinearGradient from 'react-native-linear-gradient'
+import { connect } from 'react-redux'
 
-import THEME from '../../../../theme/variables/airbitz'
+import { type EdgeTheme } from '../../../../reducers/ThemeReducer.js'
+import type { State as StateType } from '../../../../types/reduxTypes.js'
 import { getObjectDiff } from '../../../../util/utils'
 
-const REVERSE_COLORS = [THEME.COLORS.GRADIENT.DARK, THEME.COLORS.GRADIENT.LIGHT]
-const COLORS = [THEME.COLORS.GRADIENT.LIGHT, THEME.COLORS.GRADIENT.DARK]
 const UPPER_LEFT = { x: 0, y: 0 }
 const UPPER_RIGHT = { x: 1, y: 0 }
 
-export type Props = {
+export type OwnProps = {
   children?: Node,
   reverse?: boolean,
   style?: StyleSheet.Styles
 }
 
-export class Gradient extends Component<Props> {
+type StateProps = {
+  theme: EdgeTheme
+}
+
+type Props = OwnProps & StateProps
+
+class GradientComponent extends PureComponent<Props> {
   shouldComponentUpdate(nextProps: Props) {
     const diffElement = getObjectDiff(this.props, nextProps, { style: true, children: true })
     return !!diffElement
   }
 
   render() {
-    const { children, reverse, style } = this.props
+    const { children, reverse, theme, style } = this.props
+    const colors = [theme.background1, theme.background2]
+    const reverseColors = [theme.background2, theme.background1]
     return (
-      <LinearGradient style={style} start={UPPER_LEFT} end={UPPER_RIGHT} colors={reverse ? REVERSE_COLORS : COLORS}>
+      <LinearGradient style={style} start={UPPER_LEFT} end={UPPER_RIGHT} colors={reverse ? reverseColors : colors}>
         {children}
       </LinearGradient>
     )
   }
 }
 
+export const Gradient = connect((state: StateType): StateProps => ({ theme: state.theme }))(GradientComponent)
 export default Gradient
